+++ Project_QLVT/src/model/CTPNModel.java
@@ -3,7 +3,7 @@ package model;
 public class CTPNModel extends CTAbstractModel {
 	private String mapn;
 
-	public CTPNModel(String mavt, Integer soLuong, Double donGia, String mapn) {
+	public CTPNModel(String mavt, Integer soLuong, Float donGia, String mapn) {
 		super(mavt, soLuong, donGia);
 		this.mapn = mapn;
 	}
