+++ Project_QLVT/src/model/CTPXModel.java
@@ -3,7 +3,7 @@ package model;
 public class CTPXModel extends CTAbstractModel {
 	private String mapx;
 
-	public CTPXModel(String mavt, Integer soLuong, Double donGia, String mapx) {
+	public CTPXModel(String mavt, Integer soLuong, Float donGia, String mapx) {
 		super(mavt, soLuong, donGia);
 		this.mapx = mapx;
 	}
